Working on the TeamsListScreen section

+++ app/src/main/java/com/tecknobit/refy/ui/screens/TeamsListScreen.kt
@@ -2,10 +2,14 @@
 
 package com.tecknobit.refy.ui.screens
 
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -15,9 +19,13 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.GroupOff
+import androidx.compose.material.icons.filled.Share
 import androidx.compose.material3.Card
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.ListItemDefaults
+import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -28,20 +36,25 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.tecknobit.equinoxcompose.components.EmptyListUI
 import com.tecknobit.refy.R
+import com.tecknobit.refy.ui.activities.navigation.SplashScreen.Companion.user
 import com.tecknobit.refy.ui.theme.AppTypography
 import com.tecknobit.refy.ui.theme.displayFontFamily
+import com.tecknobit.refy.ui.utilities.ItemDescription
 import com.tecknobit.refy.ui.utilities.Logo
 import com.tecknobit.refy.ui.utilities.OptionsBar
+import com.tecknobit.refy.ui.utilities.RefyLinkUtilities
 import com.tecknobit.refy.ui.utilities.TeamsUtilities
+import com.tecknobit.refy.ui.utilities.getItemRelations
 import com.tecknobit.refy.ui.viewmodels.teams.TeamsListViewModel
 import com.tecknobit.refycore.records.Team
 
-class TeamsListScreen: Screen(), TeamsUtilities {
+class TeamsListScreen: Screen(), TeamsUtilities, RefyLinkUtilities {
 
     private val viewModel = TeamsListViewModel()
 
@@ -91,6 +104,7 @@ class TeamsListScreen: Screen(), TeamsUtilities {
     private fun TeamCard(
         team: Team
     ) {
+        val isMaintainer = team.isMaintainer(user)
         Card(
             modifier = Modifier
                 .fillMaxWidth()
@@ -99,9 +113,12 @@ class TeamsListScreen: Screen(), TeamsUtilities {
                     onClick = {
                         // TODO: NAV TO TEAM
                     },
-                    onLongClick = {
-                        // TODO: TO EDIT
-                    }
+                    onLongClick = if (isMaintainer) {
+                        {
+                            // TODO: TO EDIT
+                        }
+                    } else
+                        null
                 ),
             shape = RoundedCornerShape(
                 size = 8.dp
@@ -121,9 +138,15 @@ class TeamsListScreen: Screen(), TeamsUtilities {
                         team = team
                     )
                 }
-                OptionsBar(
-                    team = team
-                )
+                AnimatedVisibility(
+                    visible = isMaintainer,
+                    enter = fadeIn(),
+                    exit = fadeOut()
+                ) {
+                    OptionsBar(
+                        team = team
+                    )
+                }
             }
         }
     }
@@ -139,12 +162,24 @@ class TeamsListScreen: Screen(), TeamsUtilities {
             ),
             leadingContent = {
                 Logo(
-                    profilePicSize = 55.dp,
+                    profilePicSize = 115.dp,
                     shape = RoundedCornerShape(
                         size = 10.dp
                     ),
                     addShadow = true,
                     profilePicUrl = team.logoPic,
+                )
+            },
+            overlineContent = {
+                PicturesRow(
+                    pictures = {
+                        val profiles = mutableListOf<String>()
+                        team.members.forEach { member ->
+                            profiles.add(member.profilePic)
+                        }
+                        profiles
+                    },
+                    pictureSize = 20.dp
                 )
             },
             headlineContent = {
@@ -160,14 +195,8 @@ class TeamsListScreen: Screen(), TeamsUtilities {
                 )
             },
             supportingContent = {
-                PicturesRow(
-                    pictures = {
-                        val profiles = mutableListOf<String>()
-                        team.members.forEach { member ->
-                            profiles.add(member.profilePic)
-                        }
-                        profiles
-                    }
+                ItemDescription(
+                    description = team.description
                 )
             }
         )
@@ -178,24 +207,67 @@ class TeamsListScreen: Screen(), TeamsUtilities {
     private fun OptionsBar(
         team: Team
     ) {
-        val addToTeam = remember { mutableStateOf(false) }
+        val addLinks = remember { mutableStateOf(false) }
+        val addCollections = remember { mutableStateOf(false) }
         val deleteTeam = remember { mutableStateOf(false) }
+        val context = LocalContext.current
         OptionsBar(
             options = {
+                Row {
+                    val iconsColor = LocalContentColor.current
+                    val links = getItemRelations(
+                        userList = user.links,
+                        linkList = team.links
+                    )
+                    AddLinksButton(
+                        viewModel = viewModel,
+                        show = addLinks,
+                        links = links,
+                        team = team,
+                        tint = iconsColor
+                    )
+                    val collections = getItemRelations(
+                        userList = user.collections,
+                        linkList = team.collections
+                    )
+                    AddCollectionsButton(
+                        viewModel = viewModel,
+                        show = addCollections,
+                        collections = collections,
+                        team = team,
+                        tint = iconsColor
+                    )
+                    IconButton(
+                        onClick = {
+                            viewModel.createJoinLink(
+                                onSuccess = { joinLink ->
+                                    shareLink(
+                                        context = context,
+                                        link = joinLink
+                                    )
+                                }
+                            )
+                        }
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.Share,
+                            contentDescription = null
+                        )
+                    }
+                }
                 Column(
                     modifier = Modifier
                         .fillMaxWidth(),
                     horizontalAlignment = Alignment.End
                 ) {
-                    // TODO: TO USE CORRECTLY UNCOMMENTING THE IF AND BRACKETS
-                    //if(team.isTheAuthor(user)) {
+                    if(team.isTheAuthor(user)) {
                         DeleteTeamButton(
                             viewModel = viewModel,
                             deleteTeam = deleteTeam,
                             team = team,
                             tint = MaterialTheme.colorScheme.error
                         )
-                    //}
+                    }
                 }
             }
         )
